requirements.txt
kafka-python
pytz